Fix qreal is not double.

+++ src/plugins/valgrind/callgrindhelper.cpp
@@ -58,7 +58,7 @@ QColor CallgrindHelper::colorForString(const QString &text)
 
 QColor CallgrindHelper::colorForCostRatio(qreal ratio)
 {
-    ratio = qBound(0.0, ratio, 1.0);
+    ratio = qBound(qreal(0.0), ratio, qreal(1.0));
     return QColor::fromHsv(120 - ratio * 120, 255, 255, (-((ratio-1) * (ratio-1))) * 120 + 120);
 }
